Fixed so that it hopefully compiles without warning on _all_ platforms. Problem was that the actuall type of size_t varies from platform to platform.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@2243 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/csstring.h
@@ -80,7 +80,8 @@ public:
   {
 #ifdef DEBUG
     if (iPos > Size)
-      STR_FATAL(("Trying to access string `%s' at position %u\n", Data, iPos))
+      STR_FATAL(("Trying to access string `%s' at position %lu\n",
+        Data, (unsigned long)iPos))
 #endif
     return Data [iPos];
   }
@@ -90,7 +91,8 @@ public:
   {
 #ifdef DEBUG
     if (iPos > Size)
-      STR_FATAL (("Trying to do `%s'.SetAt (%u)\n", Data, iPos))
+      STR_FATAL (("Trying to do `%s'.SetAt (%lu)\n",
+        Data, (unsigned long)iPos))
 #endif
     Data [iPos] = iChar;
   }
@@ -100,7 +102,8 @@ public:
   {
 #ifdef DEBUG
     if (iPos > Size)
-      STR_FATAL (("Trying to do `%s'.GetAt (%u)\n", Data, iPos));
+      STR_FATAL (("Trying to do `%s'.GetAt (%lu)\n",
+        Data, (unsigned long)iPos));
 #endif
     return Data [iPos];
   }
